[Feat #8] Security 인증, 인가 예외 처리
